Drop references to the committee
No committee ended up being formed, so that part of the proposal
mistakenly gave the impression that it was possible to make a change
happen in the ecosystem by making a pull request in this repository and
waiting for "the committee" to approve it. Since there is no committee,
this leads to frustration as pull requests are never accepted.

I rewrote those parts of the proposal to emphasize the importance of
getting buy-in from the representatives of the projects which will
implement the change. There was already a paragraph about this
requirement, so this is nothing new, I merely made it more prominent in
the instructions.

The only real change I am making is to specify how a pull request gets
merged: when the author says that they consider that the proposal is
accepted, one of the maintainers of this repository will merge the pull
request. Of course, what really matters is whether the representatives
of the projects all agree to make the required changes, which is why I
emphasize this point over and over. Whether the pull request is merged
or not is only cosmetic.